// main.d.ts
export interface PockemonResults {
  name: string
  url: string
}

export interface PockemonSingleAllDetails {
abilities: PockemonAbilities [],
base_experience: number,
cries: PockemonCries,
forms: PockemonForms,
game_indices: PockemonGameIndices [],
height: number,
held_items: [],
id: number,
is_default: boolean,
location_area_encounters: string,
moves: PockemonMoves [],
name: string,
order: number,
past_abilities: [],
past_types: [],
species: PockemonSpecies [],
sprites: SpritesProps,
stats: PockemonStats [],
types: PockemonTypes [],
weight: number
}

export interface PockemonTypes {
  some: any;
  slot: number,
  type: {
    name: string,
    url: string
  }
}

export interface PockemonAbilities {
  some: any;
  ability: {
    name: string,
    url: string
  },
  is_hidden: boolean,
  slot: number
}

export interface PockemonStats {
  base_stat: number,
  effort: number,
  stat: {
    name: string,
    url: string
  }
}

export interface PockemonMoves {
  move: {
    name: string,
    url: string
  },
  version_group_details: []
}

export interface PockemonSpecies {
  name: string,
  url: string
}

export interface PockemonCries {
  latest: string,
  legacy: string
}

export interface PockemonForms {
  name: string,
  url: string
}

export interface PockemonGameIndices {
  game_index: number,
  version: {
    name: string,
    url: string
  }
}

export interface SpritesProps {
  back_default: string
  back_female: any
  back_shiny: string
  back_shiny_female: any
  front_default: string
  front_female: any
  front_shiny: string
  front_shiny_female: any
  other: OtherInOtherSprites
  versions: VersionsSprites
}
// ...
export interface IAllPokemonsAndPage {
  pokemonDetails: PockemonSingleAllDetails[] | null, page: number
}